requirements.txt
flask
Flask-MySQLdb
PyJWT
mysql-connector-python
python-dotenv
